changed db from sqlite to postgres

+++ db/migrate/20140808154119_add_total_price_to_product.rb
@@ -1,5 +1,5 @@
 class AddTotalPriceToProduct < ActiveRecord::Migration
   def change
-    add_column :products, :total_price, :double
+    add_column :products, :total_price, :decimal
   end
 end
